# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Sistem Kasir dengan Manajemen Stok dan Biaya Operasional.",
        "title": "Backend Cashier API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/api/categories": {
            "get": {
                "description": "Mengambil semua data kategori barang yang tersedia",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Data"
                ],
                "summary": "Get Daftar Kategori",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Category"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/expenditure": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Laporan Pengeluaran (Filter)",
                "responses": {}
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Input Pengeluaran (Barang/Operasional)",
                "responses": {}
            }
        },
        "/api/products": {
            "get": {
                "description": "Mengambil data produk, bisa difilter berdasarkan nama atau kode barang",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Data"
                ],
                "summary": "Get Daftar Produk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cari berdasarkan Nama atau Kode Barang",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.ProductResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/sales": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Laporan Penjualan (Filter)",
                "responses": {}
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Input Penjualan Kasir (Full Detail)",
                "parameters": [
                    {
                        "description": "Payload Penjualan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.SalesRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.SalesResponse"
                        }
                    }
                }
            }
        },
        "/master/resources": {
            "get": {
                "description": "Mengambil daftar Category, Brand, dan Supplier untuk pilihan input produk.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Data"
                ],
                "summary": "Get Dropdown Resources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Mengambil semua data produk dengan pagination dan pencarian nama/kode.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get All Products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by Name or Code",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Create New Product",
                "parameters": [
                    {
                        "description": "Product Body",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get Product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ProductRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Update Existing Product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete Product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Brand": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.Customer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "identity_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.ProductRequest": {
            "type": "object",
            "properties": {
                "brand": {
                    "$ref": "#/definitions/domain.Brand"
                },
                "brandID": {
                    "type": "integer"
                },
                "category": {
                    "$ref": "#/definitions/domain.Category"
                },
                "categoryID": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "hbeli": {
                    "type": "number"
                },
                "hjual": {
                    "type": "number"
                },
                "hpokok": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "kode": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "saldo": {
                    "type": "number"
                },
                "satuan": {
                    "type": "string"
                },
                "supplier": {
                    "$ref": "#/definitions/domain.Supplier"
                },
                "supplierID": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.ProductResponse": {
            "type": "object",
            "properties": {
                "brand": {
                    "$ref": "#/definitions/domain.Brand"
                },
                "brandID": {
                    "type": "integer"
                },
                "category": {
                    "$ref": "#/definitions/domain.Category"
                },
                "categoryID": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "hbeli": {
                    "type": "number"
                },
                "hjual": {
                    "type": "number"
                },
                "hpokok": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "kode": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "saldo": {
                    "type": "number"
                },
                "satuan": {
                    "type": "string"
                },
                "supplier": {
                    "$ref": "#/definitions/domain.Supplier"
                },
                "supplierID": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.SalesResponse": {
            "type": "object",
            "properties": {
                "amount_paid": {
                    "type": "number"
                },
                "change": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "discount": {
                    "description": "Potongan harga",
                    "type": "number"
                },
                "h_beli": {
                    "type": "number"
                },
                "h_jual": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "invoice": {
                    "type": "string"
                },
                "member_name": {
                    "description": "Dari field 'Member' di gambar",
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "product": {
                    "$ref": "#/definitions/domain.ProductResponse"
                },
                "product_id": {
                    "type": "integer"
                },
                "qty": {
                    "type": "number"
                },
                "total_netto": {
                    "description": "Harga setelah diskon",
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Supplier": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "http.SalesRequest": {
            "type": "object",
            "properties": {
                "amount_paid": {
                    "type": "number",
                    "example": 2500000
                },
                "customer": {
                    "$ref": "#/definitions/domain.Customer"
                },
                "discount": {
                    "type": "number",
                    "example": 0
                },
                "is_dp": {
                    "type": "boolean",
                    "example": false
                },
                "member_name": {
                    "type": "string",
                    "example": "Desak"
                },
                "payment_method": {
                    "type": "string",
                    "example": "Bayar Tunai"
                },
                "price": {
                    "type": "number",
                    "example": 20000
                },
                "product_search": {
                    "type": "string",
                    "example": "POLY 32RG9059"
                },
                "qty": {
                    "type": "number",
                    "example": 1
                },
                "user_id": {
                    "type": "integer",
                    "example": 123
                }
            }
        }
    }
}